Bug fix for RingBuffer class on Python < 2.7
The RingBuffer class depends on collections.deque's maxlen
property, which was only introduced in Python 2.7. Since we're
already modifying the maxlen behavior (to emit rather than
destroy the evicted element on append), it makes sense to just
implement 100% of maxlen functionality, removing the Python 2.7
dependency.

Also, raise exception if maxlen not set, and raise NotImplemented
on appendLeft(), extend(), and extendLeft(), since those have not
been overriden for proper maxlen behavior.